+++ manaba/models/ManabaQuery.py
@@ -5,33 +5,33 @@ manaba 小テスト
 import datetime
 from typing import Optional
 
-from src.models.ManabaTaskStatus import ManabaTaskStatus
+from manaba.models.ManabaTaskStatus import ManabaTaskStatus
 
 
-class ManabaReport:
+class ManabaQuery:
     """
-    manaba レポート
+    manaba 小テスト
     """
 
     def __init__(self,
-                 report_id: int,
+                 query_id: int,
                  title: str,
                  status: ManabaTaskStatus,
                  status_lamp: bool,
                  reception_start_time: Optional[datetime.datetime],
                  reception_end_time: Optional[datetime.datetime]):
         """
-        manaba レポート
+        manaba 小テスト
 
         Args:
-            report_id: レポート ID
-            title: レポートタイトル
+            query_id: 小テスト ID
+            title: 小テストタイトル
             status: ステータス
             status_lamp: ステータスランプ
-            reception_start_time: 開始日時
-            reception_end_time: 終了日時
+            reception_start_time: 受付開始日時
+            reception_end_time: 受付終了日時
         """
-        self._report_id = report_id
+        self._query_id = query_id
         self._title = title
         self._status = status
         self._status_lamp = status_lamp
@@ -39,19 +39,19 @@ class ManabaReport:
         self._reception_end_time = reception_end_time
 
     @property
-    def report_id(self) -> int:
+    def query_id(self) -> int:
         """
-        アンケート ID
+        小テスト ID
 
         Returns:
             int: ID
         """
-        return self._report_id
+        return self._query_id
 
     @property
     def title(self) -> str:
         """
-        アンケートタイトル
+        小テストタイトル
 
         Returns:
             str: タイトル
@@ -84,7 +84,7 @@ class ManabaReport:
         開始日時
 
         Returns:
-            Optional[datetime.datetime]: 開始日時
+            Optional[datetime.datetime]: 受付開始日時
         """
         return self._reception_start_time
 
@@ -94,6 +94,6 @@ class ManabaReport:
         終了日時
 
         Returns:
-            Optional[datetime.datetime]: 終了日時
+            Optional[datetime.datetime]: 受付終了日時
         """
         return self._reception_end_time
