+++ manaba/models/ManabaReportDetails.py
@@ -4,10 +4,10 @@ manaba レポート詳細
 import datetime
 from typing import Optional
 
-from src.models.ManabaPortfolioType import ManabaPortfolioType
-from src.models.ManabaResultViewType import ManabaResultViewType
-from src.models.ManabaStudentReSubmitType import ManabaStudentReSubmitType
-from src.models.ManabaTaskStatus import ManabaTaskStatus
+from manaba.models.ManabaPortfolioType import ManabaPortfolioType
+from manaba.models.ManabaResultViewType import ManabaResultViewType
+from manaba.models.ManabaStudentReSubmitType import ManabaStudentReSubmitType
+from manaba.models.ManabaTaskStatus import ManabaTaskStatus
 
 
 class ManabaReportDetails:
