+++ manaba/models/ManabaResultViewType.py
